Feat : " 스킬 데이터 구조체 ( SkillCastingTime 추가 ) "
Data
- st_SkillData SKillCastingTime 변수 타입 변경

DataManager
- SkillCastingTime 파싱 GetInt()로 변경

+++ GameServer/Data.h
@@ -85,7 +85,7 @@ struct st_SkillData
 	int32 SkillDataId;
 	string SkillName;
 	int32 SkillCoolTime;
-	float SkillCastingTime;
+	int32 SkillCastingTime;
 	int SkillDistance;
 	string SkillThumbnailImagePath;
 };
