+++ GameServer/DataManager.cpp
@@ -252,7 +252,7 @@ void CDataManager::LoadDataSkill(wstring LoadFileName)
 		int SkillDataSheetId = Filed["SkillDataSheetId"].GetInt();
 		string SkillName = Filed["SkillName"].GetString();
 		int SkillCoolTime = Filed["SkillCoolTime"].GetInt();
-		float SkillCastingTime = Filed["SkillCastingTime"].GetFloat();
+		int SkillCastingTime = Filed["SkillCastingTime"].GetInt();
 		int SkillDistance = Filed["SkillDistance"].GetInt();
 		string SkillImagePath = Filed["SkillThumbnailImagePath"].GetString();
